% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinets-package.R
\docType{package}
\name{cabinets}
\alias{cabinets}
\alias{cabinets-package}
\title{Streamlined organization with project specific custom templates}
\description{
cabinets makes it easy to create project specific file structure templates
that can be referenced at the start of any R session.
}
