% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{cabinets_options_set}
\alias{cabinets_options_set}
\title{Set cabinets options}
\usage{
cabinets_options_set(..., .envir = NULL)
}
\arguments{
\item{...}{Options to be set}

\item{.envir}{Environment to set options in; if NULL, will use environment at \code{parent.frame()}}
}
\value{
If no options are set, returns the options specified in \code{options}.
}
\description{
Set cabinets options
}
\examples{
\dontrun{
cabinets_options_set()
}
}
