% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinets.R
\name{create_cabinet}
\alias{create_cabinet}
\title{Create a cabinet template}
\usage{
create_cabinet(name, directory, structure)
}
\arguments{
\item{name}{Name of the cabinet; character of length 1. This is how the cabinet will be referenced, so best to chose something memorable.}

\item{directory}{The location your cabinet will exist.}

\item{structure}{A list of paths of folders/files to create. See details for further explanation.}
}
\description{
Create a cabinet template
}
