% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinets.R
\name{create_r_proj}
\alias{create_r_proj}
\title{R project settings}
\usage{
create_r_proj(version = "1.0", restore_workspace = c("No", "Yes",
  "Default"), save_workspace = c("No", "Yes", "Default"),
  save_history = c("Default", "No", "Yes"),
  enable_code_indexing = c("Yes", "No", "Default"),
  spaces_for_tab = c("Yes", "No", "Default"), num_spaces_for_tab = 2,
  encoding = "UTF-8", rnw_weave = "Sweave", latex = "pdfLaTeX",
  auto_append_new_line = "Yes", strip_trailing_white_space = "Yes")
}
\arguments{
\item{...}{}
}
\description{
R project settings
}
