% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{delete_cabinet}
\alias{delete_cabinet}
\title{Delete Cabinets}
\usage{
delete_cabinet(cabinet)
}
\arguments{
\item{cabinet}{An R6 object of class Cabinet written to the .Rprofile.}
}
\description{
Delete Cabinets
}
