% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinets.R
\name{edit_r_profile}
\alias{edit_r_profile}
\title{Open .Rprofile for editing}
\usage{
edit_r_profile()
}
\value{
A message that .Rprofile is being opened or that it doesn't exist.
}
\description{
\code{edit_r_profile} opens the .Rprofile file for editing.
 If the .Rprofile file doesn't exist an error message will be returned.
 This is essentially a wrapper function for \code{file.edit}.
}
