% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinets.R
\docType{data}
\name{FileCabinet}
\alias{FileCabinet}
\title{R6 object generator for class FileCabinet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileCabinet
}
\description{
R6 object generator for class FileCabinet
}
\keyword{datasets}
