% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinets.R
\name{get_cabinets}
\alias{get_cabinets}
\title{Print available cabinets}
\usage{
get_cabinets()
}
\value{
Objects of class FileCabinet found in the global environment.
}
\description{
\code{get_cabinets} returns objects of class FileCabinet found in the global environment.
}
\examples{
get_cabinets()
}
