% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinets.R
\name{new_cabinet_proj}
\alias{new_cabinet_proj}
\title{Create a new project using a cabinet template}
\usage{
new_cabinet_proj(cabinet, project_name, r_project = TRUE, open = TRUE,
  ...)
}
\arguments{
\item{cabinet}{The name of the cabinet template.}

\item{project_name}{The name of the project to store in the cabinet.}

\item{r_project}{Logical, should an Rproject be created. Default is TRUE.}

\item{open}{Logical, if creating an Rproject, should that project be opened once created. Default is TRUE.}

\item{...}{Extra arguments to pass to \code{create_r_proj}.}
}
\description{
Create a new project using a cabinet template
}
