% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{cabinets_options_set}
\alias{cabinets_options_set}
\title{Set cabinets options}
\usage{
cabinets_options_set(..., .envir = NULL)
}
\arguments{
\item{...}{Options to be set}

\item{.envir}{Environment to set options in; if NULL, will use environment at \code{parent.frame()}}
}
\value{
If no options are set, returns the options specified in \code{options}.
}
\description{
Set cabinets options
}
\details{
Mainly used for specifying if cabinets has permission to write to .Rprofile. Permission can be revoked at any time by opening the .Rprofile file and setting \code{"cabinets.permission" = FALSE}.
}
