% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinets.R
\name{FileCabinet}
\alias{FileCabinet}
\title{R6 object generator for class FileCabinet}
\description{
R6 object generator for class FileCabinet

R6 object generator for class FileCabinet
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FileCabinet$new()}}
\item \href{#method-print}{\code{FileCabinet$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileCabinet$new(name, directory, structure)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileCabinet$print(...)}\if{html}{\out{</div>}}
}

}
}
