% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabinets.R
\name{new_cabinet_proj}
\alias{new_cabinet_proj}
\title{Create a new project using a cabinet template}
\usage{
new_cabinet_proj(cabinet, project_name, r_project = TRUE, open = TRUE, ...)
}
\arguments{
\item{cabinet}{The name of the cabinet template. Available cabinets can be found using \code{get_cabinets()}.}

\item{project_name}{The name of the project to store in the cabinet, a character string.}

\item{r_project}{Logical, should an Rproject be created. Default is TRUE if working in RStudio (only works in RStudio).}

\item{open}{Logical, if creating an Rproject, should that project be opened once created. Default is TRUE if working in RStudio (only works in RStudio).}

\item{...}{Extra arguments to pass to \code{create_r_proj}.}
}
\value{
Creates a new directory at the path specified in the cabinet template. If r_project is set to TRUE, a .Rproj file will also be created using the project name. If open is set to TRUE, the new R project will opened in a new R session.
}
\description{
Generate new project directories using cabinet templates.
}
\seealso{
\code{\link{create_cabinet}}
}
